src/requirements.txt
altgraph==0.17
appJar==0.94.0
dnspython==1.16.0
macholib==1.14
more-itertools==8.0.2
mysql-connector-python==8.0.19
numpy==1.18.1
pandas==0.25.3
protobuf==3.6.1
psutil==5.6.7
pyarrow==0.15.1
PyInstaller==3.6
python-dateutil==2.8.1
pytz==2019.3
six==1.13.0
